Removed availableTasks configuration property
availableTasks can be modeled as a within-subject factor. For a single
repeatable task A, the task factor will have only one level A. For
non-repeatable tasks A, B and C, the task factor will have three
levels A, B and C. If we can assume that tasks do not have a
carry-over effect, then the task factor can be counterbalanced with a
sequential order type. On the other hand, if we expect the task
factor to have symmetrical carry-over effect, then we can apply
counterbalancing.